requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
PyCookieCloud==1.0.3
pycryptodomex==3.20.0
pyTelegramBotAPI==4.22.1
PyYAML==6.0.2
requests==2.32.3
schedule==1.2.2
telebot==0.0.5
urllib3==2.2.2
